Add files via upload
I received the problem to solve for the lowest positive integer not included in an array. I’m sure that there are many ways to solve this problem, but I wanted to find a solution that compared values of a variable to existing values in array. To do this I made a loop to create a new array that only included positive integers, and then sorted the array from lowest to highest. This throws out data that is not needed, and puts it into a format that makes the answer potentially easy to see if printed out for limited testing. I created an initial value of 1 and tested to see if that value was in the array. If the value of my array equates to the index value than 1 is added to the index and it continues. This goes on until the index value represents an integer that is not included in the initial array. Because the index value starts at 1, the first value not included in the array represents the lowest positive integer value. 

This was the problem question. 

This is a demo task.
Write a function:
def solution(A)
that, given an array A of N integers, returns the smallest positive integer (greater than 0) that does not occur in A.
For example, given A = [1, 3, 6, 4, 1, 2], the function should return 5.
Given A = [1, 2, 3], the function should return 4.
Given A = [−1, −3], the function should return 1.
Write an efficient algorithm for the following assumptions:
N is an integer within the range [1..100,000];
each element of array A is an integer within the range [−1,000,000..1,000,000].